fix(price): convert datatype double to decimal

+++ eTickets.Business/Interfaces/IShoppingCartService.cs
@@ -5,9 +5,9 @@ namespace eTickets.Business.Interfaces
     public interface IShoppingCartService
     {
         IReadOnlyList<ShoppingCartItem> GetItems();
-        void AddToCart(int movieId, string movieName, string movieImageUrl, double price, int quantity = 1);
+        void AddToCart(int movieId, string movieName, string movieImageUrl, decimal price, int quantity = 1);
         void RemoveFromCart(int movieId, int quantity = 1);
-        double GetTotal();
+        decimal GetTotal();
         int GetCount();
         void ClearCart();
     }
